♻️ youtube表示のコンポーネントを分離
