+++ frontend/app/vue-app/src/components/ToDoItem.vue
@@ -4,6 +4,7 @@ import Timer from "./Timer";
 import alarm from "@/assets/sound-alarm.mp3";
 import requestAPI from "./requestAPI";
 import { useRoute } from "vue-router";
+import YouTube from "./YouTube.vue";
 
 const defaultTaskTime = 25;
 const defaultBreakTime = 5;
@@ -23,7 +24,6 @@ let totalPassedTime;
 const circumference = 2 * Math.PI * 45;
 const remainingTime = ref(0);
 const isMenuOpen = ref(false);
-const youtubeUrl = ref("");
 const routeId = useRoute().params.id;
 
 const error = ref(null);
@@ -126,26 +126,6 @@ const stop = () => {
   timer.setTime(remainingTime.value);
 };
 
-const getYouTubeId = (url) => {
-  const regExp =
-    /^(https:\/\/www\.youtube\.com\/watch\?v=|https:\/\/youtu\.be\/)([a-zA-Z0-9-_]+)(\?|$)/;
-  const match = url.match(regExp);
-  return match ? match[2] : null;
-};
-
-const roadYouTube = () => {
-  const youtubeId = getYouTubeId(youtubeUrl.value);
-  youtubeUrl.value = "";
-  if (youtubeId) {
-    const iframe = document.querySelector("iframe");
-    iframe.width = "640";
-    iframe.height = "360";
-    iframe.src = `https://www.youtube.com/embed/${youtubeId}`;
-  } else {
-    alert("無効なURLです");
-  }
-};
-
 const customizeTimer = () => {
   if (timer.getIsTimerRunning()) {
     timer.reset();
@@ -190,22 +170,7 @@ const customizeTimer = () => {
         />
       </button>
     </div>
-    <div v-show="isMenuOpen" class="video-container">
-      <div class="video-input-form">
-        <input type="text" v-model="youtubeUrl" placeholder="YouTube URL" />
-        <button @click="roadYouTube()">ロード</button>
-      </div>
-      <iframe
-        width="0"
-        height="0"
-        src=""
-        title="YouTube video player"
-        frameborder="0"
-        allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share"
-        referrerpolicy="strict-origin-when-cross-origin"
-        allowfullscreen
-      ></iframe>
-    </div>
+    <YouTube v-show="isMenuOpen" />
     <div class="timer-customize">
       <div class="time-selector">
         <label for="task-time">タスク時間</label>
@@ -340,42 +305,6 @@ const customizeTimer = () => {
   font-family: "Lucida Console", monospace;
   margin: 8px 8px 8px 64px;
   padding: 8px;
-}
-
-.video-container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin: 1rem auto;
-  width: 55%;
-}
-
-.video-input-form {
-  display: flex;
-  justify-content: center;
-  width: 50%;
-  margin: 0rem auto 1rem auto;
-}
-
-.video-input-form input {
-  width: 60%;
-  padding: 0.5rem;
-  border: 1px solid #ccc;
-  border-radius: 4px;
-  margin: 0 0.5rem 0 0;
-  font-size: 14px;
-}
-
-.video-input-form button {
-  padding: 0.5rem 1rem;
-  margin: 0 0 0 0.5rem;
-  border: 1px solid black;
-  border-radius: 4px;
-  background-color: red;
-  color: white;
-  cursor: pointer;
-  font-size: 14px;
-  font-weight: bold;
 }
 
 .timer-customize {
